requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
openai==1.3.5
requests==2.31.0
pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6